Refactoring and testing

+++ structs.h
@@ -7,6 +7,8 @@ struct studentfs_dirp {
 struct metadata {
 	char curr_vnum[MAX_VNUM_LEN];
 	size_t vcount;    // # of checkpoints
-	size_t vmax;      // Maximum number of checkpoints
-	size_t size_freq; // Checkpointing frequency based on size of change
+
+	/* For configuring checkpointing. -1 indicates indefinite */
+	int vmax;         // Maximum number of checkpoints
+	int size_freq;    // Checkpointing frequency based on size of change
 };
